Bill S5528 version original - James Skoufis
BILL NUMBER: S5528 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the state law, in relation to the description of the arms of the state and the state flag PURPOSE OR GENERAL IDEA OF BILL: The purpose of this legislation is to restore the New York State coat of arms and state flag to its original state. SUMMARY OF SPECIFIC PROVISIONS: Section one: amends section 70 of the state law by removing the line 'E pluribus unum.