Bill S5534 version original - William Weber
BILL NUMBER: S5534 SPONSOR: WEBER TITLE OF BILL: An act to amend the education law, in relation to adjusting the regional cost index for state school aid with respect to the Hudson Valley region PURPOSE: Adjusts the regional labor force cost index for the Hudson Valley region. SUMMARY OF PROVISIONS: Section 1: Amends the education law to adjust the regional cost index in the Hudson Valley.