Bill S5532 version original - Rachel May
BILL NUMBER: S5532 SPONSOR: MAY TITLE OF BILL: An act to amend the executive law, in relation to establishing the LGBTQ+ advisory board PURPOSE: To establish a permanent advisory board on LGBTQ+ issues. SUMMARY OF PROVISIONS: Section 1: Sets forth the make-up of and selection for the LGBTQ+ Advi- sory Board and length of terms of service.