Bill S5579A version original - Jessica Scarcella-Spanton
BILL NUMBER: S5579 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the insurance law, in relation to requiring insurance policies to provide coverage for transvaginal ultrasounds during preg- nancy PURPOSE: The purpose of this bill is to ensure that testing for an incompetent cervix is covered by insurance. SUMMARY OF PROVISIONS: Section 1.