Bill S5569 version original - Jabari Brisport
BILL NUMBER: S5569 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the criminal procedure law, in relation to the filing of pre-trial motions in criminal cases SUMMARY OF PROVISIONS: Section one amends subdivision 1 of section 255.20 of the criminal procedure 2 law, with respect to the timing when pre-trial motions may be filed. JUSTIFICATION: This amendment to the criminal procedure law will allow a defendant to file pre-trial motions after having the opportunity to review complete discovery.