Bill S5571A version original - Andrew Gounardes
BILL NUMBER: S5571 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend chapter 540 of the laws of 2021 amending the retirement and social security law relating to determination of salary base for members of the city of New York fire department pension fund, in relation to the application thereof PURPOSE OR GENERAL IDEA OF BILL: To make Chapter 540 of 2021 retroactive to the beginning of Tier 2. SUMMARY OF PROVISIONS: Section one of this bill changes the applicability of the effective date of Ch.