Bill S5602 version original - Leroy Comrie
BILL NUMBER: S5602 SPONSOR: COMRIE TITLE OF BILL: An act to amend the agriculture and markets law, in relation to creating an advisory board on food safety protection and inspection programs in the department of agriculture and markets PURPOSE OF BILL: To improve State regulation of food safety. SUMMARY OF SPECIFIC PROVISIONS: Creates a thirteen Member advisory board within the Department of Agri- culture and Markets on food safety and inspection programs.