Bill S5599 version original - Leroy Comrie
BILL NUMBER: S5599 SPONSOR: COMRIE TITLE OF BILL: An act to amend the correction law, in relation to an annual report of department statistics SUMMARY OF PROVISIONS: This legislation requires the department of corrections and community supervision to provide an annual inmate mortality report. Further,. it requires the department to provide a report to the legislation with seven days following the suicide or other death of an inmate.