Bill S5544 version original - Leroy Comrie
BILL NUMBER: S5544 SPONSOR: COMRIE TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to requiring the chair of the state liquor authority to make certain quar- terly reports PURPOSE OR GENERAL IDEA OF BILL: Relates to requiring the chairman of the state liquor authority to make certain quarterly reports. SUMMARY OF SPECIFIC PROVISIONS: Section 1.