Bill S5559 version original - Leroy Comrie
BILL NUMBER: S5559 SPONSOR: COMRIE TITLE OF BILL: An act to amend the public service law and the public authorities law, in relation to providing for a two percent cap on rate increases imposed by utilities PURPOSE: To better regulate and contain cost burden on utility ratepayers in New York State. SUMMARY OF PROVISIONS: Subdivision 1 of section 65 of the public service law, as amended by chapter 789 of the laws of 1930, is amended as described in the bill.