Bill S5620B version original - Pete Harckham
BILL NUMBER: S5620 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the executive law, in relation to adding Doodletown Brook to the definition of inland waterways for the purposes of water- front revitalization PURPOSE OR GENERAL IDEA OF BILL: Designates the Doodletown brook as an inland waterway. SUMMARY OF PROVISIONS: Amends the executive law, in relation to adding Doodletown Brook to the definition of inland waterways for the purposes of waterfront revitali- zation.