Bill S5650A version original - John Liu
BILL NUMBER: S5650 SPONSOR: LIU TITLE OF BILL: An act to amend the education law, in relation to expanding access to advanced courses to improve educational equity PURPOSE: This bill seeks to expand access to advanced coursework in order to improve educational equity around the state. SUMMARY OF PROVISIONS: This bill creates a new section of education law that provides for notice to parents on the availability of advanced courses, and other early college and career experiences.