Bill S5637 version original - Jake Ashby
BILL NUMBER: S5637 SPONSOR: ASHBY TITLE OF BILL: An act to amend chapter 22 of the laws of 2014, relating to expanding opportunities for service-disabled veteran-owned business enterprises, in relation to making certain provisions thereof permanent PURPOSE OR GENERAL IDEA OF BILL: Makes permanent certain provisions of chapter 22 of the laws of 2014, relating to expanding opportunities for service-disabled veteran-owned business enterprises. SUMMARY OF PROVISIONS: Section 1.