Bill S5641A version original - Leroy Comrie
BILL NUMBER: S5641 SPONSOR: COMRIE TITLE OF BILL: An act to amend the labor law, in relation to establishing criteria for the use of automated employment decision tools PURPOSE OR GENERAL IDEA OF BILL: To limit bias and discrimination against any group on the basis of sex, race, ethnicity, or other protected class that occurs when automated employment decision tools are used to screen candidates for hire. SUMMARY OF PROVISIONS: Section 1 amends the labor law by adding a new section 203-f.