Bill S5690 version original - Kevin S. Parker
BILL NUMBER: S5690 SPONSOR: PARKER TITLE OF BILL: An act to amend the legislative law, in relation to racial and ethnic impact statements on bills PURPOSE OR GENERAL IDEA OF BILL: This bill would require an impact statement attached to qualifying bills. Qualifying bills are legislation that can have adverse disparate effects on minority groups. SUMMARY OF PROVISIONS: Section 1: This act shall be known and may be cited as the "racial and ethnic equity act".