Bill S5686 version original - Monica Martinez
BILL NUMBER: S5686 SPONSOR: MARTINEZ TITLE OF BILL: An act to repeal article 23 of the tax law, in relation to the metropol- itan commuter transportation mobility tax PURPOSE: The purpose of this legislation is to repeal the the Metropolitan Commu- ter Transportation Mobility Tax. SUMMARY OF PROVISIONS: Section 1. repeals article 23 of the tax law known as the Metropolitan Commuter Transportation Mobility Tax Section 2.