Bill S5670 version original - Patrick M. Gallivan
BILL NUMBER: S5670 SPONSOR: GALLIVAN TITLE OF BILL: An act to amend the parks, recreation and historic preservation law, in relation to the use of snowmobiles in the town of Concord, county of Erie PURPOSE: This bill permits snowmobile use along a portion of the State Route 219 right-of-way in the Town of Concord, Erie County, upon designation by the New York State Department of Transportation and the Office of Parks, Recreation and Historic Preservation. SUMMARY OF PROVISIONS: Section 1.