Bill S5700 version original - Kevin S. Parker
BILL NUMBER: S5700 SPONSOR: PARKER TITLE OF BILL: An act to amend the social services law, in relation to written notifi- cation to schools of a student's placement in foster care PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to notify the Principal or designated school official of the initial school enrollment of a child in foster care. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The social services law is amended by adding a new section 405.