Bill S5673 version original - Nathalia Fernandez
BILL NUMBER: S5673 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the tax law, in relation to providing a tax credit for small businesses employing a person previously convicted of a crime PURPOSE OR GENERAL IDEA OF BILL: This legislation will provide small businesses with a tax credit for hiring non-violent felons. SUMMARY OF PROVISIONS: Section 1: Section of the 210-B of the tax law is amended to establish the rehabilitation opportunity tax credit.