Bill S5702 version original - Kevin S. Parker
BILL NUMBER: S5702 SPONSOR: PARKER TITLE OF BILL: An act to amend the education law, in relation to prohibiting the sale of food and food products containing boneless lean beef trimmings, commonly referred to as "pink slime", in schools PURPOSE: Prohibits the sale of food and food products containing boneless lean beef trimmings, commonly referred to as "pink slime", in schools. SUMMARY OF PROVISIONS: Section 1: The education law is amended by adding a new section 915-a.