Bill S5662 version original - Andrew Gounardes
BILL NUMBER: S5662 SPONSOR: GOUNARDES TITLE OF BILL: An act to amend the executive law and the tax law, in relation to estab- lishing the data economy labor compensation and accountability act PURPOSE OR GENERAL IDEA OF BILL: To establish the Office of Consumer Data Protection to better regulate companies who profit off of consumer data and to create a tax on their gross receipts SUMMARY OF PROVISIONS: Section one of this bill creates a short title. Section two establishes legislative intent.