Bill S5687A version original - Monica Martinez
BILL NUMBER: S5687 SPONSOR: MARTINEZ TITLE OF BILL: An act to amend the public authorities law, in relation to freezing the salary of the executives in the metropolitan transportation authority PURPOSE: The purpose of this bill improve the Metropolitan Transportation Author- ity's services. SUMMARY OF PROVISIONS: Section 1. provides a title for the act "MTA Executive Pay Freeze" Section 2. is the legislative intent Section 3.