Bill S5666 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5666 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the general business law, in relation to the sale and use of employment information PURPOSE OR GENERAL IDEA OF BILL: Bans the sale of employment information, including payroll and earnings information, hours worked, consumer history and health insurance infor- mation. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 380-v of the article is re-lettered section 380-w, and a new section 380-v is added.