Bill S5665 version original - Timothy M. Kennedy
BILL NUMBER: S5665 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the penal law, in relation to criminal impersonation of a disabled veteran vendor PURPOSE: This bill designates criminal impersonation of a disabled veteran vendor as a class A misdemeanor. SUMMARY OF PROVISIONS: Section 1: Amends section 190.25 of the penal law to expand the defi- nition of second degree criminal impersonation to include the impersona- tion of a specially licensed disabled veteran vendor.