Bill S5708 version original - Timothy M. Kennedy
BILL NUMBER: S5708 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the penal law, in relation to financial exploitation of the elderly and disabled PURPOSE: This bill will allow prosecutions for financial abuse of senior citizens by amending the Penal Law to add a new crime of financial exploitation of the elderly or disabled. SUMMARY OF SPECIFIC PROVISIONS: This bill amends section 155.