Bill S5698 version original - Kevin S. Parker
BILL NUMBER: S5698 SPONSOR: PARKER TITLE OF BILL: An act to amend the workers' compensation law, in relation to enacting the "protection in the workplace act" PURPOSE OR GENERAL IDEA OF BILL: To clarify that the workers' compensation law was never intended to bar employees suffering a personal injury as a result of a sexual assault in the workplace from pursuing legal action seeking damages from their employer. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Short title.