Bill S5685 version original - Monica Martinez
BILL NUMBER: S5685 SPONSOR: MARTINEZ TITLE OF BILL: An act in relation to appointing a fiscal monitor for the metropolitan transportation authority; and providing for the repeal of such provisions upon expiration thereof PURPOSE: The purpose of this bill is to improve the fiscal stability of the MTA through accountability and review. SUMMARY OF PROVISIONS: Section 1. provides definitions Section 2.