Bill S5669 version original - James Sanders Jr.
BILL NUMBER: S5669 SPONSOR: SANDERS TITLE OF BILL: An act in relation to the study of the utilization of state government payments as an opportunity to transition unbanked and underbanked state residents into the banking system PURPOSE: This bill would require a study by the Department of Financial Services of the utilization of state government payments as an opportunity to transition unbanked and underbanked state residents into the banking. SUMMARY OF PROVISIONS: Section 1.