Bill S5660 version original - James Sanders Jr.
BILL NUMBER: S5660 SPONSOR: SANDERS TITLE OF BILL: An act to amend the public service law, in relation to the community wind energy pilot program PURPOSE: The bill shall establish a New York state pilot program promoting the use of community wind energy. The commission shall oversee the implemen- tation of the pilot program and shall issue an order of the commission authorizing the construction of a wind energy system installation.