Bill S5770 version original - Kevin S. Parker
BILL NUMBER: S5770 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law and the state finance law, in relation to enacting the "Empire State Windfall Profits Tax Act of 2023" PURPOSE OR GENERAL IDEA OF BILL: To amend the tax law and the state finance law, in relation to enacting the "Empire State Windfall Profits Tax Act of 2023" SUMMARY OF PROVISIONS: Section 1: Short title. This act shall be known and may be cited as the "Empire State Windfall Profits Tax Act of 2023".