Bill S5731 version original - James Skoufis
BILL NUMBER: S5731 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the alcoholic beverage control law, in relation to permitting the sale of beer at retail on Sundays; and to repeal certain provisions of the alcoholic beverage control law relating thereto PURPOSE: This legislation would permit beer to be sold for off-premise consump- tion between the hours of 3AM and SAM on Sundays. SUMMARY OF PROVISIONS: Section 1: Repeals section 105-a of the alcoholic beverage control law.