Bill S5782A version original - George Borrello
BILL NUMBER: S5782 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the executive law, in relation to including Silver Lake Outlet within the definition of "inland waterways" for the purposes of waterfront revitalization PURPOSE: To included Silver Lake Outlet within the definition of "inland water- ways" for the purpose of waterfront realization. SUMMARY OF SPECFIC PROVISIONS: Section 1. Paragraph (a) consists of the state's major inland lakes, with Silver Lake Outlet (in Wyoming).