Bill S5733 version original - George Borrello
BILL NUMBER: S5733 SPONSOR: BORRELLO TITLE OF BILL: An act to authorize the board of education of the Salamanca city school district to establish a federal impact aid reserve fund in the event federal impact aid is reduced PURPOSE OR GENERAL IDEA OF BILL: Authorizes the City School District of the City of Salamanca to estab- lish a reserve fund for federal impact aid in the event that such aid is reduced. SUMMARY OF PROVISIONS: Section 1. Provides Legislative Findings. Section 2.