Bill S5742 version original - Anthony H. Palumbo
BILL NUMBER: S5742 SPONSOR: PALUMBO TITLE OF BILL: An act permitting the village board of the village of Southampton to provide health insurance coverage for enrolled members of the Southamp- ton Village Ocean Rescue PURPOSE: This legislation recognizes the work and commitment of the Southampton Village Ocean Rescue by polluting the village of Southampton to extend health insurance coverage to enrolled members. SUMMARY OF PROVISIONS: Section 1.