Bill S5723 version original - Timothy M. Kennedy
BILL NUMBER: S5723 SPONSOR: KENNEDY TITLE OF BILL: An act to amend the insurance law, in relation to allowing insurance carriers to offer a discount to drivers who have been trained to operate an authorized emergency vehicle PURPOSE OR GENERAL IDEA OF BILL: The purpose of this bill is to allow insurance carriers to offer a discount to drivers who have been trained to operate an authorized emer- gency vehicle. SUMMARY OF PROVISIONS: Section 1.