Bill S5762 version original - Kevin S. Parker
BILL NUMBER: S5762 SPONSOR: PARKER TITLE OF BILL: An act to amend the insurance law, in relation to establishing an advi- sory panel on health insurance coverage for autism spectrum disorder PURPOSE: To establish an advisory panel to report to the superintendent regarding successful treatment options for autism spectrum disorder that will be required to be covered by insurance. SUMMARY OF PROVISIONS: Section 1: The insurance law is amended by adding a new section 346.