Bill S5757 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5757 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the executive law, in relation to the establishment of an insurance liaison on the disaster preparedness commission PURPOSE: The purpose of this bill is to provide the position of insurance liaison on the disaster preparedness commission. SUMMARY OF PROVISIONS: Section 1 amends the executive law to allow for the appointment of an expert insurance liaison on the disaster preparedness committee.