Bill S5737 version original - Joseph A. Griffo
BILL NUMBER: S5737 SPONSOR: GRIFFO TITLE OF BILL: An act to amend the vehicle and traffic law, in relation to allowing snowmobiles and all-terrain vehicles to be registered as farm vehicles PURPOSE: The purpose of this legislation is to allow for snowmobiles and all-ter- rain vehicles to be registered as a farm vehicle for legitimate farming purposes. SUMMARY OF PROVISIONS: Section 1 allows for the registration of a snowmobile or all-terrain vehicle as a farm vehicle.