Bill S5756 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5756 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the education law, in relation to the appointment of a military liaison to each school district SUMMARY OF PROVISIONS: Section 1. The commissioner shall direct that each school district, charter school, private elementary, and secondary school appoint a mili- tary liaison for purposed of providing military families with informa- tion regarding benefits they may be eligible or under the laws of this state.