Bill S5754 version original - Thomas F. O'Mara
BILL NUMBER: S5754 SPONSOR: O'MARA TITLE OF BILL: An act to amend the penal law, in relation to authorizing any justice or judge in the state to obtain a license to have and carry concealed a pistol or revolver PURPOSE OR GENERAL IDEA OF BILL: Authorizes any justice or judge in the state to obtain a license to have and carry concealed a pistol or revolver. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Paragraph (d) of subdivision 2 of section 400.