Bill S5750 version original - Kevin S. Parker
BILL NUMBER: S5750 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to creating the office of nutrition and fitness PURPOSE OR GENERAL IDEA OF BILL:. To create an office of nutrition and fitness in order to address the issue of obesity in children and adults. SUMMARY OF SPECIFIC PROVISIONS: Section 1: The executive law is amended by adding a new article 24-A. Section 2: This act shall take effect on the one hundred eightieth day after it shall have become a law.