Bill S5740 version original - Anthony H. Palumbo
BILL NUMBER: S5740 SPONSOR: PALUMBO TITLE OF BILL: An act to amend the penal law, in relation to creating the crime of unlawful possession of a radio frequency jamming device PURPOSE: Creates the crime of unlawful possession of a radio frequency jamming device. SUMMARY OF PROVISIONS: Section 1 amends the penal law by adding a new section 270.45 to read as follows: Unlawful possession of a radio frequency jamming device.