Bill S5776 version original - Kevin S. Parker
BILL NUMBER: S5776 SPONSOR: PARKER TITLE OF BILL: An act to amend the public service law, the public authorities law and the environmental conservation law, in relation to enacting the "coal to gas plant conversion and redevelopment blueprint act" PURPOSE: Enacts the "Coal to Gas Plant Conversion and Redevelopment Act" SUMMARY OF SPECIFIC PROVISIONS: Section 1: Short title. This act shall be known and may be cited as the "coal to gas plant conversion and redevelopment blueprint act".