Bill S5773 version original - Kevin S. Parker
BILL NUMBER: S5773 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to establishing a workgroup to conduct analysis on the ambulatory patient group rates and commercial insurance rates for behavioral health services PURPOSE: The purpose of this bill is to ensure that mental health and substance use service providers have fiscally viable programs. SUMMARY OF PROVISIONS: Section 1: Section 7.07 of the mental hygiene law is amended by adding a new subdivision (i).