Bill S5755 version original - George Borrello
BILL NUMBER: S5755 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the tax law, in relation to establishing a fine reimbursement credit for taxpayers fined for operating a bar or restau- rant during the COVID-19 pandemic PURPOSE OR GENERAL IDEA OF BILL: Establishes a fine reimbursement credit for taxpayers fined for operat- ing a bar or restaurant during the COVID-19 pandemic SUMMARY OF PROVISIONS: Section 1. Section 606 of the tax law is amended by adding a new subsection (kkk).