Bill S5767 version original - Kevin S. Parker
BILL NUMBER: S5767 SPONSOR: PARKER TITLE OF BILL: An act to amend the executive law, in relation to providing for awards of attorney's fees in actions under the human rights law PURPOSE: To provide access to justice to those who cannot afford it in actions or proceedings relating to the state human rights law. SUMMARY OF PROVISIONS: Section 1: Subdivision 10 of section 297 of the executive law, as amended by chapter 154 of the laws of 2022, is amended.