Bill S5777 version original - James Skoufis
BILL NUMBER: S5777 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the public health law, in relation to extending wage parity for home care workers in upstate New York PURPOSE: This bill would extend the wage parity requirement upstate and require a wage of $20.42 as of October 1. SUMMARY OF PROVISIONS: Section 1 - Extends the wage parity requirement upstate and requires a wage of $20.42. Section 2 - Effective date.