Bill S5710 version original - Mark Walczyk
BILL NUMBER: S5710 SPONSOR: WALCZYK TITLE OF BILL: An act to amend the state finance law, in relation to requiring state agencies to expend funds no later than ninety days after receipt of a certificate of approval from the director of the budget PURPOSE OR GENERAL IDEA OF BILL: This bill will require state agencies to pay grants and other monies awarded in a timely manner. If they do not, the agency will be forced to forfeit some part of their budget the following fiscal year.