Bill S5753 version original - Kevin S. Parker
BILL NUMBER: S5753 SPONSOR: PARKER TITLE OF BILL: An act to amend the insurance law, in relation to requiring notification to insured persons that an out-of-network physician may be used in their procedure, test or surgery PURPOSE OR GENERAL IDEA OF BILL: To provide notice to patients that services provided may not be covered by health insurance. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 3216 of the insurance law is amended by adding a new subsection (n).