Bill S5797B version original - Nathalia Fernandez
BILL NUMBER: S5797 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the criminal procedure law, in relation to the privacy and confidentiality of urine test results performed pursuant to a judi- cial diversion program PURPOSE OR GENERAL IDEA OF BILL: Provides that the results of urinalysis performed on an individual participating in a judicial diversion program will be confidential and immune to criminal action. SUMMARY OF PROVISIONS: Section 1 amends subdivision 5 of section 216.