Bill S5795A version original - Dean Murray
BILL NUMBER: S5795 SPONSOR: MURRAY TITLE OF BILL: An act in relation to authorizing the assessor of the town of Brookhaven to accept an application for exemption from real property taxes from Al-Muneer Foundation, Inc. PURPOSE: To Authorize the assessor of the town of Brookhaven to accept an appli- cation for exemption for real property taxes with respect to 2019-2020 and 2020-2021 assessment rolls from Al-Muneer Foundation, Inc.