Bill S5787 version original - Julia Salazar
BILL NUMBER: S5787 SPONSOR: SALAZAR TITLE OF BILL: An act to amend the education law, in relation to the unauthorized or unlicensed practice of massage therapy PURPOSE: The purpose of this legislation is to decriminalize unlicensed massage work. SUMMARY OF PROVISIONS: Section 1 amends Section 6512 subdivision 1 of the education law to add a new subdivision, which stipulates that unauthorized or unlicensed practice of massage therapy is not a violation of this section.