Bill S5786 version original - Julia Salazar
BILL NUMBER: S5786 SPONSOR: SALAZAR TITLE OF BILL: An act to amend the criminal procedure law and the family court act, in relation to statements of those accused of crimes to enhance criminal investigations and prosecutions and to promote confidence in the crimi- nal justice system of this state PURPOSE: To enhance the integrity of the criminal legal system and to prevent coerced or false confessions. SUMMARY OF PROVISIONS: Section 1. Amends subdivision 3 of section 60.