Bill S5801A version original - John Liu
BILL NUMBER: S5801 SPONSOR: LIU TITLE OF BILL: An act to amend the public officers law, in relation to the award of reasonable attorneys' fees in certain proceedings PURPOSE: To Provide for the award of reasonable attorneys'. fees in FOIL proceedings if the person is successful and in open meeting proceedings to the successful party. SUMMARY OF PROVISIONS: Section 1.