Bill S5798 version original - Nathalia Fernandez
BILL NUMBER: S5798 SPONSOR: FERNANDEZ TITLE OF BILL: An act to amend the civil rights law, in relation to the right to breast feed PURPOSE OR GENERAL IDEA OF BILL: To change the verbiage from "mother" to "person" in regards to the right to breastfeed. SUMMARY OF PROVISIONS: Section 1 amends the civil rights law by changing the term "mother" to "person" and using gender neutral pronouns. Section 2 sets the effective date. JUSTIFICATION: Words matter.