Bill S5811A version original - George Borrello
BILL NUMBER: S5811 SPONSOR: BORRELLO TITLE OF BILL: An act to amend the tax law, in relation to extending the expiration of the provisions authorizing the county of Allegany to impose an addi- tional one and one-half percent sales and compensating use taxes PURPOSE: To extend the expiration of provisions authorizing the county of Allega- ny to impose an additional one and one-half percent sales and compensat- ing use taxes. SUMMARY OF PROVISIONS: Section 1.