Bill S5789 version original - James Sanders Jr.
BILL NUMBER: S5789 SPONSOR: SANDERS TITLE OF BILL: An act to amend the civil rights law, in relation to discrimination in the ownership of cooperative housing PURPOSE OR GENERAL IDEA OF BILL: Many prospective cooperators are arbitrarily and discriminatorily rejected. By requiring cooperative apartment corporations, to provide a statement of reasons, this bill will serve to reduce the incidence of such rejections.