Bill S5813 version original - Rachel May
BILL NUMBER: S5813 SPONSOR: MAY TITLE OF BILL: An act to amend the environmental conservation law, in relation to declaring the waters of the state as public trust resources PURPOSE OR GENERAL IDEA OF BILL: This bill declares the waters of the State to be public trust resources. SUMMARY OF PROVISIONS: This bill amends § 15-0103 of the environmental conservation law by adding a new subdivision 13 to declare that all waters of the state are deemed to be public trust resources.