Bill S5831A version original - Robert Jackson
BILL NUMBER: S5831 SPONSOR: JACKSON TITLE OF BILL: An act to amend the retirement and social security law, in relation to the disability retirement for safety and security officers PURPOSE: To provide safety and security officers with a 3/4 disability retirement benefit. SUMMARY OF PROVISIONS: Section 1 amends Section 607-a of the Retirement and Social Security Law to provide safety and security officers with a 3/4 disability retirement benefit. Section 2 is the effective date.