Bill S5840 version original - James Sanders Jr.
BILL NUMBER: S5840 SPONSOR: SANDERS TITLE OF BILL: An act to amend the real property law, in relation to establishing a paired testing program and a compliance fee; and to amend the state finance law, in relation to establishing the paired testing program fund PURPOSE OF THE BILL: The purpose of this bill is to establish a paired testing program and a compliance fee to prevent housing discrimination. SUMMARY OF PROVISIONS: Section 1: Definition. Paired testing program.