Bill S5846 version original - James Sanders Jr.
BILL NUMBER: S5846 SPONSOR: SANDERS TITLE OF BILL: An act to amend the tax law, in relation to establishing the work oppor- tunity tax credit PURPOSE OR GENERAL IDEA OF BILL: To provide a state tax credit for small businesses in order to incentiv- ize the hiring of long term unemployed persons. SUMMARY OF PROVISIONS: Section one of the bill amends section 210-B of the tax law by adding a new subdivision 49, which establishes the work opportunity tax credit.