Bill S5834 version original - James Skoufis
BILL NUMBER: S5834 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the highway law, in relation to designating a portion of the state highway system as the "Catholic War Veterans Memorial Highway" PURPOSE OR GENERAL IDEA OF BILL: Relates to the designation of the "Catholic War Veterans Memorial High- way". SUMMARY OF PROVISIONS: Section 1: Amends the highway law by adding a new section 344-o to designate a portion of the state highway system as the "Catholic War Veterans Memorial Highway".