Bill S5835 version original - Luis R. Sepúlveda
BILL NUMBER: S5835 SPONSOR: SEPULVEDA TITLE OF BILL: An act to amend the surrogate's court procedure act, in relation to compensation for deputy public administrators in the city of New York PURPOSE: Requires that each of the deputy public administrators of the counties compromising the city of New York shall receive as compensation at least two-thirds of the amount paid or hereafter paid to the public adminis- trators of each county, respectively. SUMMARY OF PROVISIONS: Section 1.