Bill S5865 version original - Dean Murray
BILL NUMBER: S5865 SPONSOR: MURRAY TITLE OF BILL: An act in relation to authorizing the assessor of the town of Brookha- ven, county of Suffolk, to accept from the Post-Morrow Foundation, Inc. an application for exemption from real property taxes PURPOSE: The purpose of the bill is to authorize the assessor of the town of Brookhaven, county of Suffolk, to accept from the Post-Morrow Founda- tion, Inc. an application for exemption from real property taxes.