Bill S5845 version original - James Sanders Jr.
BILL NUMBER: S5845 SPONSOR: SANDERS TITLE OF BILL: An act to amend the education law, the state finance law, the social services law, the correction law, the administrative code of the city of New York and the civil service law, in relation to high school equiv- alency diplomas PURPOSE: To ensure there is no financial hurdle for high school equivalency diplomas. SUMMARY OF PROVISIONS: This bill prohibits a fee for admission to any exam that would award a high school equivalency diploma.