Bill S5864 version original - Kevin Thomas
BILL NUMBER: S5864 SPONSOR: THOMAS TITLE OF BILL: An act to amend the education law, in relation to career and technical education courses provided by boards of cooperative educational services PURPOSE: To establish equity of opportunity for students to enroll in career and technical education programs. SUMMARY OF PROVISIONS: Section 1 of the bill provides the legislative findings.