Bill S5880 version original - James Skoufis
BILL NUMBER: S5880 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the vehicle and traffic law and the New York city char- ter, in relation to permitted use of handicapped spaces in the city of New York PURPOSE: This bill provides that vehicles with permits for handicapped parking issued anywhere in the state may use all spaces for handicapped parking in the city of New York. The bill also provides that handicapped parking permits issued by the City of New York will be valid outside of the City.