Bill S5858 version original - Kevin S. Parker
BILL NUMBER: S5858 SPONSOR: PARKER TITLE OF BILL: An act to amend the mental hygiene law, in relation to a public aware- ness educational program on heroin; and making an appropriation therefor PURPOSE OR GENERAL IDEA OF BILL: To direct the Commissioner of the Office of Alcohol and Substance Abuse to create a public awareness and educational program on the dangers and costs of heroin use, and allocate the funds necessary for such program. SUMMARY OF SPECIFIC PROVISIONS: Section 1: Section 19.