Bill S5859 version original - Kevin S. Parker
BILL NUMBER: S5859 SPONSOR: PARKER TITLE OF BILL: An act to amend the tax law, in relation to providing a tax credit for certain taxpayers who donate blood PURPOSE OF THE BILL: Establishes a personal income tax credit for taxpayers who donate blood to a blood bank four or more times in a year. SUMMARY OF PROVISIONS: Section 1: Section 606 of the tax law is amended by adding a new subsection (000).