Bill S5878 version original - Kristen Gonzalez
BILL NUMBER: S5878 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the tax law, in relation to establishing a tax on carbon emissions from non-essential helicopter flights in cities with a popu- lation of one million or more PURPOSE OR GENERAL IDEA OF BILL: This bill will impose a carbon tax and a noise tax on non-essential helicopter flights in New York City. The revenue from these taxes will be dedicated to and administered by the State's Environmental Protection Fund.