Bill S4266A version A - Lea Webb
BILL NUMBER: S4266A SPONSOR: WEBB TITLE OF BILL: An act to establish a task force on missing women and girls who are black, indigenous and people of color; and providing for the repeal of such provisions upon expiration thereof PURPOSE: To develop a task force on Missing BIPOC Women and Girls that will work to address the discrepancies in lack of care/concern for missing and murdered Black, Indigenous, women and girls of color. SUMMARY OF PROVISIONS: Section 1 sets forth legislative findings.