Bill S5881A version original - James Skoufis
BILL NUMBER: S5881 SPONSOR: SKOUFIS TITLE OF BILL: An act to amend the real property law, in relation to manufactured home parks PURPOSE OR GENERAL IDEA OF BILL: This legislation relates to the sale of manufactured home parks. SUMMARY OF SPECIFIC PROVISIONS: Section 1. Amends Section 233-a of the Real Property Law relating to the sale of manufactured home parks. Section 2. Provides for an effective date of 180 days after becoming law.