Bill S5886 version original - Pete Harckham
BILL NUMBER: S5886 SPONSOR: HARCKHAM TITLE OF BILL: An act to amend the village law, in relation to exempting the Buchanan Engine Company No. 1 from the requirement that the percentage of non-re- sident fire department members not exceed forty-five percent of the membership PURPOSE: Exempts the Buchanan Engine Company No.