Bill S5885 version original - Jessica Scarcella-Spanton
BILL NUMBER: S5885REVISED 3/27/23 SPONSOR: SCARCELLA-SPANTON TITLE OF BILL: An act to amend the criminal procedure law, in relation to the discovery of material, non-material and impeachment information and the required time frame for such discovery SUMMARY OF SPECIFIC PROVISIONS: Section 1. Subdivision 4 of section 30.30 of the criminal procedure law is amended as follows: 4.