Bill S5915 version original - Kristen Gonzalez
BILL NUMBER: S5915 SPONSOR: GONZALEZ TITLE OF BILL: An act to amend the executive law, in relation to the provision of informational materials on economic abuse PURPOSE: This legislation would require the creation and distribution of informa- tional materials on economic abuse and the resources available to those in need. SUMMARY OF SPECIFIC PROVISIONS: Section one amends the executive law by adding a new section 76. Section two establishes the effective date.