Bill S5898 version original - Jabari Brisport
BILL NUMBER: S5898 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the environmental conservation law, in relation to prohibiting the administration of psychoactive drugs to an animal in a zoo in certain situations SUMMARY OF PROVISIONS: Section one amends the environmental conservation law by adding section 11-0541, which prohibits the use of psychoactive drugs on animals in zoos for the purposes of procreation. Section two sets the effective date.