Bill S5902 version original - Jabari Brisport
BILL NUMBER: S5902 SPONSOR: BRISPORT TITLE OF BILL: An act to amend the penal law, in relation to requiring proof of liabil- ity insurance prior to the issuance of a license to carry a firearm SUMMARY OF PROVISIONS: Section one of the bill amends subdivision 1 of § 400.00 of the penal law by adding the requirement of proof of personal liability insurance prior to the issuance or renewal of a firearm license. Section two of the bill provides the effective date.